Remove object and array decorating for now
What: Remove decorating {} and [] for now and leaving rules for how
to decorate strings and plain values.

Why: My hypothesis is that the array and string decoration and
formatting should be done in the functions that handle them separately
and the decoratePrimitives() can be applied inside them when there is
no more nesting detected.

Just a hunch tho idk.