Add stringifyObject() to stringifyJSON() functionality
What: Add partially complete stringifyObject() abilities (flat objects
and nested objects but not objects with nested arrays) to the main
program stringifyJSON().
Why: Bringing it together gradually into the main program to be tested.

What: Replace the mocha tests from specific helpers to the main app as
each helper is included into the main app.
Why: So I know that the helpers are working in the context of the main
app and not just as independent helpers.